调整命名规则；将要修改比赛在客户端进行，从服务器进行剥离；

+++ PokemenEXEServer/RandomGenerator.hpp
@@ -14,7 +14,7 @@ namespace Pokemen
 		std::random_device rd;
 		g_Rand_Engine.seed(rd());
 
-		std::uniform_int_distribution<> u(0, limit - 1);
+		std::uniform_int_distribution<> u(0, (int)limit - 1);
 		return u(g_Rand_Engine);
 	}
